requirements.txt
flask
stripe
python-dotenv
pymysql
cryptography
flask-dance[google] 
